+++ lib/models/account_data.dart
@@ -1,21 +1,22 @@
-import 'package:meta/meta.dart';
+import 'package:equatable/equatable.dart';
 import 'package:sacco/models/transactions/export.dart';
 
-class AccountData {
+class AccountData extends Equatable {
   final String accountNumber;
   final String sequence;
   final List<StdCoin> coins;
 
-  AccountData({
-    @required this.accountNumber,
-    @required this.sequence,
-    @required this.coins,
-  })  : assert(accountNumber != null),
-        assert(sequence != null),
-        assert(coins != null);
+  const AccountData({
+    required this.accountNumber,
+    required this.sequence,
+    required this.coins,
+  });
 
   @override
   String toString() {
-    return "number: $accountNumber, sequence: $sequence, coins: $coins";
+    return 'number: $accountNumber, sequence: $sequence, coins: $coins';
   }
+
+  @override
+  List<Object?> get props => [accountNumber, sequence, coins];
 }