+++ lib/models/export.dart
@@ -1,4 +1,3 @@
 export 'account_data.dart';
-export 'node_info.dart';
 export 'transaction_result.dart';
 export 'transactions/export.dart';