+++ lib/models/transaction_result.dart
@@ -1,11 +1,6 @@
 import 'package:equatable/equatable.dart';
-import 'package:json_annotation/json_annotation.dart';
-import 'package:meta/meta.dart';
-
-part 'transaction_result.g.dart';
 
 /// Represents the result that is returned when broadcasting a transaction.
-@JsonSerializable(explicitToJson: true)
 class TransactionResult extends Equatable {
   /// String representing the hash of the transaction.
   /// Note that this hash is always present, even if the transaction was
@@ -19,45 +14,55 @@ class TransactionResult extends Equatable {
   /// Please note that this field is going to be:
   /// - `null` if [success] is `true`.
   /// - a valid [TransactionError] if [success] is `false`
-  final TransactionError error;
+  final TransactionError? error;
 
-  TransactionResult({
-    @required this.hash,
-    @required this.success,
+  const TransactionResult({
+    required this.hash,
+    required this.success,
     this.error,
-  })  : assert(hash != null),
-        assert(success || error != null);
+  }) : assert(success || error != null);
 
   @override
-  List<Object> get props {
-    return [hash, success, error];
-  }
+  List<Object?> get props => [hash, success, error];
 
   factory TransactionResult.fromJson(Map<String, dynamic> json) =>
-      _$TransactionResultFromJson(json);
+      TransactionResult(
+        hash: json['hash'] as String,
+        success: json['success'] as bool,
+        error: json['error'] == null
+            ? null
+            : TransactionError.fromJson(json['error']),
+      );
 
-  Map<String, dynamic> toJson() => _$TransactionResultToJson(this);
+  Map<String, dynamic> toJson() => {
+        'hash': hash,
+        'success': success,
+        'error': error?.toJson(),
+      };
 }
 
 /// Contains the data related to an error that has occurred when
 /// broadcasting the transaction.
-@JsonSerializable(explicitToJson: true)
 class TransactionError extends Equatable {
   final int errorCode;
   final String errorMessage;
 
-  TransactionError({
-    @required this.errorCode,
-    @required this.errorMessage,
+  const TransactionError({
+    required this.errorCode,
+    required this.errorMessage,
   });
 
   @override
-  List<Object> get props {
-    return [errorCode, errorMessage];
-  }
+  List<Object?> get props => [errorCode, errorMessage];
 
   factory TransactionError.fromJson(Map<String, dynamic> json) =>
-      _$TransactionErrorFromJson(json);
+      TransactionError(
+        errorCode: json['errorCode'] as int,
+        errorMessage: json['errorMessage'] as String,
+      );
 
-  Map<String, dynamic> toJson() => _$TransactionErrorToJson(this);
+  Map<String, dynamic> toJson() => {
+        'errorCode': errorCode,
+        'errorMessage': errorMessage,
+      };
 }