+++ lib/models/transactions/std_coin.dart
@@ -1,15 +1,11 @@
-import 'package:meta/meta.dart';
+import 'package:equatable/equatable.dart';
 
 /// Contains the data of a specific coin
-class StdCoin {
+class StdCoin extends Equatable {
   final String denom;
   final String amount;
 
-  const StdCoin({
-    @required this.denom,
-    @required this.amount,
-  })  : assert(denom != null),
-        assert(amount != null);
+  const StdCoin({required this.denom, required this.amount});
 
   factory StdCoin.fromJson(Map<String, dynamic> json) => StdCoin(
         denom: json['denom'] as String,
@@ -17,7 +13,10 @@ class StdCoin {
       );
 
   Map<String, dynamic> toJson() => {
-        'denom': this.denom,
-        'amount': this.amount,
+        'denom': denom,
+        'amount': amount,
       };
+
+  @override
+  List<Object?> get props => [denom, amount];
 }