+++ lib/models/transactions/std_fee.dart
@@ -1,17 +1,17 @@
-import 'package:meta/meta.dart';
+import 'package:equatable/equatable.dart';
 import 'package:sacco/models/transactions/std_coin.dart';
 
-class StdFee {
+class StdFee extends Equatable {
   final String gas;
   final List<StdCoin> amount;
 
-  const StdFee({
-    @required this.amount,
-    @required this.gas,
-  }) : assert(gas != null);
+  const StdFee({required this.amount, required this.gas});
 
   Map<String, dynamic> toJson() => {
-        'amount': this.amount.map((coin) => coin.toJson()).toList(),
-        'gas': this.gas,
+        'amount': amount.map((coin) => coin.toJson()).toList(),
+        'gas': gas,
       };
+
+  @override
+  List<Object?> get props => [gas, amount];
 }