+++ lib/models/transactions/std_msg.dart
@@ -1,20 +1,16 @@
-import 'package:meta/meta.dart';
+import 'package:equatable/equatable.dart';
 
 /// [StdMsg] represents a standard message that can be included inside
 /// a transaction.
-class StdMsg {
+class StdMsg extends Equatable {
   /// String representing the type of the message.
   final String type;
 
   /// Map containing the real value of the message.
-  Map<String, dynamic> value;
+  final Map<String, dynamic> value;
 
   /// Public constructor.
-  StdMsg({
-    @required this.type,
-    @required this.value,
-  })  : assert(type != null),
-        assert(value != null);
+  const StdMsg({required this.type, required this.value});
 
   /// Converts this instance of [StdMsg] into a map that can be later used
   /// to serialize it as a JSON object.
@@ -22,4 +18,7 @@ class StdMsg {
         'type': type,
         'value': value,
       };
+
+  @override
+  List<Object?> get props => [type, value];
 }