+++ lib/tx_signer.dart
@@ -1,8 +1,9 @@
 import 'dart:convert';
 
-import 'package:meta/meta.dart';
+import 'package:sacco/models/node_info.dart';
 import 'package:sacco/sacco.dart';
 import 'package:sacco/utils/export.dart';
+import 'package:http/http.dart' as http;
 
 import 'utils/map_sorter.dart';
 
@@ -12,16 +13,31 @@ class TxSigner {
   /// given [wallet] and returns a new [StdTx] containing the signatures
   /// inside it.
   static Future<StdTx> signStdTx({
-    @required Wallet wallet,
-    @required StdTx stdTx,
+    required Wallet wallet,
+    required StdTx stdTx,
+    http.Client? client,
   }) async {
+    client ??= http.Client();
+
     // Get the account data and node info from the network
-    final account = await AccountDataRetrieval.getAccountData(wallet);
-    final nodeInfo = await NodeInfoRetrieval.getNodeInfo(wallet);
+    final account = await AccountDataRetrieval.getAccountData(
+      wallet,
+      client: client,
+    );
+    final nodeInfo = await NodeInfoRetrieval.getNodeInfo(
+      wallet,
+      client: client,
+    );
 
     // Sign all messages
     final signatures = _getStdSignature(
-        wallet, account, nodeInfo, stdTx.messages, stdTx.fee, stdTx.memo);
+      wallet,
+      account,
+      nodeInfo,
+      stdTx.messages,
+      stdTx.fee,
+      stdTx.memo,
+    );
 
     // Assemble the transaction
     return StdTx(
@@ -55,20 +71,20 @@ class TxSigner {
     final sortedJson = MapSorter.sort(jsonSignature);
 
     // Encode the sorted JSON to a string and get the bytes
-    var jsonData = json.encode(sortedJson);
+    final jsonData = jsonEncode(sortedJson);
     final bytes = utf8.encode(jsonData);
 
     // Sign the data
     final signatureData = wallet.signTxData(bytes);
 
     // Get the compressed Base64 public key
-    final pubKeyCompressed = wallet.ecPublicKey.Q.getEncoded(true);
+    final pubKeyCompressed = wallet.ecPublicKey.Q!.getEncoded(true);
 
     // Build the StdSignature
     return StdSignature(
       value: base64Encode(signatureData),
       publicKey: StdPublicKey(
-        type: "tendermint/PubKeySecp256k1",
+        type: 'tendermint/PubKeySecp256k1',
         value: base64Encode(pubKeyCompressed),
       ),
     );