+++ lib/wallet.dart
@@ -1,15 +1,15 @@
 import 'dart:math';
 import 'dart:typed_data';
 
-import 'package:bip32/bip32.dart' as bip32;
-import 'package:bip39/bip39.dart' as bip39;
+import 'package:convert/convert.dart';
 import 'package:equatable/equatable.dart';
-import 'package:hex/hex.dart';
-import 'package:meta/meta.dart';
 import 'package:pointycastle/export.dart';
-import 'package:pointycastle/src/utils.dart' as pc_utils;
 import 'package:sacco/sacco.dart';
 import 'package:sacco/utils/bech32_encoder.dart';
+import 'package:sacco/utils/big_int_big_endian.dart';
+import 'package:sacco/utils/bip32.dart';
+import 'package:sacco/utils/bip39.dart';
+import 'package:sacco/utils/ecc_secp256k1.dart';
 
 import 'utils/tx_signer.dart';
 
@@ -28,14 +28,12 @@ class Wallet extends Equatable {
 
   final NetworkInfo networkInfo;
 
-  Wallet({
-    @required this.networkInfo,
-    @required this.address,
-    @required this.privateKey,
-    @required this.publicKey,
-  })  : assert(networkInfo != null),
-        assert(privateKey != null),
-        assert(publicKey != null);
+  const Wallet({
+    required this.networkInfo,
+    required this.address,
+    required this.privateKey,
+    required this.publicKey,
+  });
 
   @override
   List<Object> get props {
@@ -49,38 +47,38 @@ class Wallet extends Equatable {
   factory Wallet.derive(
     List<String> mnemonic,
     NetworkInfo networkInfo, {
-    String lastDerivationPathSegment = "0",
+    String lastDerivationPathSegment = '0',
   }) {
     // Get the mnemonic as a string
     final mnemonicString = mnemonic.join(' ');
-    if (!bip39.validateMnemonic(mnemonicString)) {
-      throw Exception("Invalid mnemonic " + mnemonicString);
+    if (!Bip39.validateMnemonic(mnemonicString)) {
+      throw Exception('Invalid mnemonic ' + mnemonicString);
     }
 
     final _lastDerivationPathSegmentCheck =
         int.tryParse(lastDerivationPathSegment) ?? -1;
     if (_lastDerivationPathSegmentCheck < 0) {
-      throw Exception("Invalid index format ${lastDerivationPathSegment}");
+      throw Exception('Invalid index format $lastDerivationPathSegment');
     }
 
     // Convert the mnemonic to a BIP32 instance
-    final seed = bip39.mnemonicToSeed(mnemonicString);
-    final root = bip32.BIP32.fromSeed(seed);
+    final seed = Bip39.mnemonicToSeed(mnemonicString);
+    final root = Bip32.fromSeed(seed);
 
     // Get the node from the derivation path
     final derivedNode =
-        root.derivePath("$BASE_DERIVATION_PATH/$lastDerivationPathSegment");
+        root.derivePath('$BASE_DERIVATION_PATH/$lastDerivationPathSegment');
 
     // Get the curve data
-    final secp256k1 = ECCurve_secp256k1();
+    final secp256k1 = ECCSecp256k1();
     final point = secp256k1.G;
 
     // Compute the curve point associated to the private key
-    final bigInt = BigInt.parse(HEX.encode(derivedNode.privateKey), radix: 16);
+    final bigInt = BigInt.parse(hex.encode(derivedNode.privateKey!), radix: 16);
     final curvePoint = point * bigInt;
 
     // Get the public key
-    final publicKeyBytes = curvePoint.getEncoded();
+    final publicKeyBytes = curvePoint!.getEncoded();
 
     // Get the address
     final sha256Digest = SHA256Digest().process(publicKeyBytes);
@@ -90,7 +88,7 @@ class Wallet extends Equatable {
     return Wallet(
       address: address,
       publicKey: publicKeyBytes,
-      privateKey: derivedNode.privateKey,
+      privateKey: derivedNode.privateKey!,
       networkInfo: networkInfo,
     );
   }
@@ -113,29 +111,30 @@ class Wallet extends Equatable {
   /// Returns the associated [publicKey] as a Bech32 string
   String get bech32PublicKey {
     final type = [235, 90, 233, 135, 33]; // "addwnpep"
-    final prefix = networkInfo.bech32Hrp + "pub";
+    final prefix = networkInfo.bech32Hrp + 'pub';
     final fullPublicKey = Uint8List.fromList(type + publicKey);
     return Bech32Encoder.encode(prefix, fullPublicKey);
   }
 
   /// Returns the associated [privateKey] as an [ECPrivateKey] instance.
   ECPrivateKey get _ecPrivateKey {
-    final privateKeyInt = BigInt.parse(HEX.encode(privateKey), radix: 16);
-    return ECPrivateKey(privateKeyInt, ECCurve_secp256k1());
+    final privateKeyInt = BigInt.parse(hex.encode(privateKey), radix: 16);
+    return ECPrivateKey(privateKeyInt, ECCSecp256k1());
   }
 
   /// Returns the associated [publicKey] as an [ECPublicKey] instance.
   ECPublicKey get ecPublicKey {
-    final secp256k1 = ECCurve_secp256k1();
+    final secp256k1 = ECCSecp256k1();
     final point = secp256k1.G;
     final curvePoint = point * _ecPrivateKey.d;
-    return ECPublicKey(curvePoint, ECCurve_secp256k1());
+    return ECPublicKey(curvePoint, ECCSecp256k1());
   }
 
   /// Signs the given [data] using the associated [privateKey] and encodes
   /// the signature bytes to be included inside a transaction.
-  Uint8List signTxData(Uint8List data) {
-    final hash = SHA256Digest().process(data);
+  Uint8List signTxData(List<int> data) {
+    final uint8List = Uint8List.fromList(data);
+    final hash = SHA256Digest().process(uint8List);
     return TransactionSigner.deriveFrom(hash, _ecPrivateKey, ecPublicKey);
   }
 
@@ -153,8 +152,8 @@ class Wallet extends Equatable {
   /// More details at
   /// https://github.com/web3j/web3j/blob/master/crypto/src/main/java/org/web3j/crypto/ECDSASignature.java#L27
   static ECSignature _toCanonicalised(ECSignature signature) {
-    final ECDomainParameters _params = ECCurve_secp256k1();
-    final BigInt _halfCurveOrder = _params.n >> 1;
+    final ECDomainParameters _params = ECCSecp256k1();
+    final _halfCurveOrder = _params.n >> 1;
     if (signature.s.compareTo(_halfCurveOrder) > 0) {
       final canonicalisedS = _params.n - signature.s;
       signature = ECSignature(signature.r, canonicalisedS);
@@ -165,27 +164,32 @@ class Wallet extends Equatable {
   /// Signs the given [data] using the private key associated with this wallet,
   /// returning the signature bytes ASN.1 DER encoded.
   Uint8List sign(Uint8List data) {
-    final ecdsaSigner = Signer("SHA-256/ECDSA")
+    final ecdsaSigner = Signer('SHA-256/ECDSA')
       ..init(
           true,
           ParametersWithRandom(
             PrivateKeyParameter(_ecPrivateKey),
             _getSecureRandom(),
           ));
-    ECSignature ecSignature =
-        _toCanonicalised(ecdsaSigner.generateSignature(data));
-    final sigBytes = Uint8List.fromList(
-      pc_utils.encodeBigInt(ecSignature.r) +
-          pc_utils.encodeBigInt(ecSignature.s),
-    );
+    final ecSignature =
+        _toCanonicalised(ecdsaSigner.generateSignature(data) as ECSignature);
+
+    const bigIntEndian = BigIntBigEndian();
+    final sigBytes = Uint8List.fromList(bigIntEndian.encode(ecSignature.r) +
+        bigIntEndian.encode(ecSignature.s));
+
     return sigBytes;
   }
 
   /// Creates a new [Wallet] instance from the given [json] and [privateKey].
   factory Wallet.fromJson(Map<String, dynamic> json, Uint8List privateKey) {
+    final address =
+        Uint8List.fromList(hex.decode(json['hex_address'] as String));
+    final publicKey =
+        Uint8List.fromList(hex.decode(json['public_key'] as String));
     return Wallet(
-      address: HEX.decode(json["hex_address"] as String),
-      publicKey: HEX.decode(json['public_key'] as String),
+      address: address,
+      publicKey: publicKey,
       privateKey: privateKey,
       networkInfo: NetworkInfo.fromJson(json['network_info']),
     );
@@ -194,9 +198,9 @@ class Wallet extends Equatable {
   /// Converts the current [Wallet] instance into a JSON object.
   /// Note that the private key is not serialized for safety reasons.
   Map<String, dynamic> toJson() => <String, dynamic>{
-        'hex_address': HEX.encode(this.address),
-        'bech32_address': this.bech32Address,
-        'public_key': HEX.encode(this.publicKey),
-        'network_info': this.networkInfo.toJson(),
+        'hex_address': hex.encode(address),
+        'bech32_address': bech32Address,
+        'public_key': hex.encode(publicKey),
+        'network_info': networkInfo.toJson(),
       };
 }