+++ test/network_info_test.dart
@@ -4,10 +4,10 @@ import 'package:test/test.dart';
 void main() {
   test('toJson and fromJson work properly with optional fields', () {
     final networkInfo = NetworkInfo(
-      bech32Hrp: "bech32",
-      lcdUrl: "lcd-url",
-      iconUrl: "icon-url",
-      name: "name",
+      bech32Hrp: 'bech32',
+      lcdUrl: Uri.parse('http://lcd-url'),
+      iconUrl: 'icon-url',
+      name: 'name',
     );
 
     final json = networkInfo.toJson();